requirements.txt
# Core Dependencies
streamlit>=1.32.0,<2.0.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.4,<2.0.0
python-dotenv>=1.0.1,<2.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0
supabase>=2.0.0,<3.0.0

# Authentication
bcrypt>=4.1.2,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.6,<1.0.0

# PDF Generation
fpdf2>=2.7.8,<3.0.0
reportlab>=4.0.0,<5.0.0

# Data Processing
openpyxl>=3.1.3,<4.0.0
matplotlib>=3.8.3,<4.0.0

# Utilities
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3.post1,<2025.0
typing-extensions>=4.8.0,<5.0.0
